[Bronze V] Title: 이상한 기호, Time: 120 ms, Memory: 14156 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -12,8 +12,8 @@ public class Main {
 
         StringTokenizer st = new StringTokenizer(reader.readLine());
 
-        int a = Integer.parseInt(st.nextToken());
-        int b = Integer.parseInt(st.nextToken());
+        long a = Integer.parseInt(st.nextToken());
+        long b = Integer.parseInt(st.nextToken());
 
         System.out.println((a + b) * (a - b));
